+++ foo_ui_columns/pch.h
@@ -55,6 +55,7 @@
 #include <wil/cppwinrt.h>
 #include <wil/com.h>
 #include <wil/resource.h>
+#include <wil/win32_helpers.h>
 
 #include <winrt/windows.foundation.h>
 #include <winrt/windows.ui.viewmanagement.h>