+++ foo_ui_columns/status_bar.h
@@ -34,7 +34,6 @@ void set_menu_item_description(std::string_view text);
 void clear_menu_item_description();
 void create_window();
 void destroy_window();
-void on_status_font_change();
 std::optional<LRESULT> handle_draw_item(LPDRAWITEMSTRUCT lpdis);
 
 } // namespace cui::status_bar